Fix color parameters (again)

+++ django_app/tuckbox/tasks.py
@@ -22,7 +22,7 @@ def build_box(self, parameters):
     my_box.create_box_file(parameters['filename'], progress_tracker)
 
     for face in parameters['faces']:
-        if parameters['faces'][face][:2] != '0x':
+        if parameters['faces'][face][:2] != '0x' and parameters['faces'][face][0] != '#':
             os.remove(parameters['faces'][face])
     # TODO: Should set a timeout timer here to forget the tasks / delete the file from here
